- Show current count of Radicoida Unica scans - Support zoom in/out on PlotGrounded - Don't render green bio scans over red ones - Add Dodec's as a supported build type

+++ SrvSurvey/plotters/PlotBase2.cs
@@ -552,7 +552,7 @@ namespace SrvSurvey.plotters
         /// <summary>
         /// A manually set scaling factor given by users with the `z` command.
         /// </summary>
-        public float customScale = -1.0f;
+        public double customScale = -1.0d;
 
         protected decimal radius { get => game.systemBody?.radius ?? 0; }
